+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 # Changelog
 
-## Next
+## 0.5.0 -- 2026-05-26
 
 * Fix: site-name of docs does not follow renaming
 * Extract flatten_latex to its own package [latex-include](https://pypi.org/project/latex-include/)